Feat swagger

+++ Api/Models/ProductModel.cs
@@ -10,15 +10,15 @@ namespace value_analysis_server.Api.Models
         public string code { get; set; }
         public string name { get; set; }
         public string description { get; set; }
-        public int quantity { get; set; }
+        public float quantity { get; set; }
         [AllowNull]
-        public int? provider1 { get; set; }
+        public float? provider1 { get; set; }
         [AllowNull]
-        public int? provider2 { get; set; }
+        public float? provider2 { get; set; }
         [AllowNull]
-        public int? provider3 { get; set; }
+        public float? provider3 { get; set; }
         [AllowNull]
-        public int? provider4 { get; set; }
+        public float? provider4 { get; set; }
         [AllowNull]
         public string? selectedProvider { get; set; }
         public DateTime created_at { get; set; }
